Modularize ext sync tester config
