+++ op-acceptance-tests/tests/sync_tester/sync_tester_ext_el/sync_tester_ext_el_test.go
@@ -22,64 +22,7 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 )
 
-// Configuration defaults for op-sepolia
-const (
-	DefaultNetworkPreset = "op-sepolia"
-
-	// Tailscale networking endpoints
-	DefaultL2ELEndpointTailscale       = "https://proxyd-l2-sepolia.primary.client.dev.oplabs.cloud"
-	DefaultL1CLBeaconEndpointTailscale = "https://beacon-api-proxy-sepolia.primary.client.dev.oplabs.cloud"
-	DefaultL1ELEndpointTailscale       = "https://proxyd-l1-sepolia.primary.client.dev.oplabs.cloud"
-)
-
-var (
-	// Network presets for different networks against which we test op-node syncing
-	networkPresets = map[string]stack.ExtNetworkConfig{
-		"op-sepolia": {
-			L2NetworkName:      "op-sepolia",
-			L1ChainID:          eth.ChainIDFromUInt64(11155111),
-			L2ELEndpoint:       "https://ci-sepolia-l2.optimism.io",
-			L1CLBeaconEndpoint: "https://ci-sepolia-beacon.optimism.io",
-			L1ELEndpoint:       "https://ci-sepolia-l1.optimism.io",
-		},
-		"base-sepolia": {
-			L2NetworkName:      "base-sepolia",
-			L1ChainID:          eth.ChainIDFromUInt64(11155111),
-			L2ELEndpoint:       "https://base-sepolia-rpc.optimism.io",
-			L1CLBeaconEndpoint: "https://ci-sepolia-beacon.optimism.io",
-			L1ELEndpoint:       "https://ci-sepolia-l1.optimism.io",
-		},
-		"unichain-sepolia": {
-			L2NetworkName:      "unichain-sepolia",
-			L1ChainID:          eth.ChainIDFromUInt64(11155111),
-			L2ELEndpoint:       "https://unichain-sepolia-rpc.optimism.io",
-			L1CLBeaconEndpoint: "https://ci-sepolia-beacon.optimism.io",
-			L1ELEndpoint:       "https://ci-sepolia-l1.optimism.io",
-		},
-		"op-mainnet": {
-			L2NetworkName:      "op-mainnet",
-			L1ChainID:          eth.ChainIDFromUInt64(1),
-			L2ELEndpoint:       "https://op-mainnet-rpc.optimism.io",
-			L1CLBeaconEndpoint: "https://ci-mainnet-beacon.optimism.io",
-			L1ELEndpoint:       "https://ci-mainnet-l1.optimism.io",
-		},
-		"base-mainnet": {
-			L2NetworkName:      "base-mainnet",
-			L1ChainID:          eth.ChainIDFromUInt64(1),
-			L2ELEndpoint:       "https://base-mainnet-rpc.optimism.io",
-			L1CLBeaconEndpoint: "https://ci-mainnet-beacon.optimism.io",
-			L1ELEndpoint:       "https://ci-mainnet-l1.optimism.io",
-		},
-	}
-	L2CLSyncMode = getSyncMode("L2_CL_SYNCMODE")
-)
-
-func getSyncMode(envVar string) sync.Mode {
-	if value := os.Getenv(envVar); value == sync.ELSyncString {
-		return sync.ELSync
-	}
-	return sync.CLSync
-}
+var L2CLSyncMode = getSyncMode("L2_CL_SYNCMODE")
 
 func TestSyncTesterExtEL(gt *testing.T) {
 	t := devtest.SerialT(gt)
@@ -161,22 +104,8 @@ func setupOrchestrator(gt *testing.T, t devtest.T, blocksToSync uint64) (*sysgo.
 	ctx := t.Ctx()
 	require := t.Require()
 
-	config := networkPresets[DefaultNetworkPreset]
-
-	// Override configuration with Tailscale endpoints if Tailscale networking is enabled
-	if os.Getenv("TAILSCALE_NETWORKING") == "true" {
-		config.L2ELEndpoint = getEnvOrDefault("L2_EL_ENDPOINT_TAILSCALE", DefaultL2ELEndpointTailscale)
-		config.L1CLBeaconEndpoint = getEnvOrDefault("L1_CL_BEACON_ENDPOINT_TAILSCALE", DefaultL1CLBeaconEndpointTailscale)
-		config.L1ELEndpoint = getEnvOrDefault("L1_EL_ENDPOINT_TAILSCALE", DefaultL1ELEndpointTailscale)
-	}
-
-	if os.Getenv("NETWORK_PRESET") != "" {
-		var ok bool
-		config, ok = networkPresets[os.Getenv("NETWORK_PRESET")]
-		if !ok {
-			gt.Errorf("NETWORK_PRESET %s not found", os.Getenv("NETWORK_PRESET"))
-		}
-	}
+	config, err := GetNetworkPreset(os.Getenv("NETWORK_PRESET"))
+	require.NoError(err, "failed to initialize network preset")
 
 	// Runtime configuration values
 	l.Info("Runtime configuration values for TestSyncTesterExtEL")
@@ -246,10 +175,9 @@ func setupOrchestrator(gt *testing.T, t devtest.T, blocksToSync uint64) (*sysgo.
 	return orch.(*sysgo.Orchestrator), target
 }
 
-// getEnvOrDefault returns the environment variable value or the default if not set
-func getEnvOrDefault(envVar, defaultValue string) string {
-	if value := os.Getenv(envVar); value != "" {
-		return value
+func getSyncMode(envVar string) sync.Mode {
+	if value := os.Getenv(envVar); value == sync.ELSyncString {
+		return sync.ELSync
 	}
-	return defaultValue
+	return sync.CLSync
 }
